+++ server/src/graql/gremlin/sets/ValueFragmentSet.java
@@ -19,25 +19,24 @@
 package grakn.core.graql.gremlin.sets;
 
 import com.google.common.collect.ImmutableSet;
-import grakn.core.graql.executor.property.ValueExecutor;
+import grakn.core.graql.executor.property.value.Operation;
 import grakn.core.graql.gremlin.EquivalentFragmentSet;
 import grakn.core.graql.gremlin.fragment.Fragment;
 import grakn.core.graql.gremlin.fragment.Fragments;
 import graql.lang.property.VarProperty;
 import graql.lang.statement.Variable;
-
-import javax.annotation.Nullable;
 import java.util.Objects;
 import java.util.Set;
+import javax.annotation.Nullable;
 
 class ValueFragmentSet extends EquivalentFragmentSet {
 
     private final Variable var;
-    private final ValueExecutor.Operation<?, ?> operation;
+    private final Operation<?, ?> operation;
 
     @Nullable private final VarProperty varProperty;
 
-    ValueFragmentSet(@Nullable VarProperty varProperty, Variable var, ValueExecutor.Operation<?, ?> operation) {
+    ValueFragmentSet(@Nullable VarProperty varProperty, Variable var, Operation<?, ?> operation) {
         this.varProperty = varProperty;
         if (var == null) {
             throw new NullPointerException("Null var");
@@ -58,7 +57,7 @@ class ValueFragmentSet extends EquivalentFragmentSet {
         return varProperty;
     }
 
-    ValueExecutor.Operation<?, ?> operation() {
+    Operation<?, ?> operation() {
         return operation;
     }
